django/requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-camel-case==1.0.3
docutils==0.14
drf-yasg==1.16.0
future==0.18.2
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.15.0
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
oauthlib==3.1.0
Pillow==6.2.0
psycopg2==2.8.4
python3-openid==3.1.0
pytz==2019.2
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
